refactor: move shared helpers that aren’t gentle-specific from gentle/ into util/
