+++ gentle/transcription.py
@@ -97,7 +97,7 @@ class Transcription:
 if __name__=='__main__':
     # full transcription
     from Queue import Queue
-    from gentle import ffmpeg
+    from util import ffmpeg
     from gentle import standard_kaldi
     import tempfile
